Projects Data Science/Adversarial Attack (1st project)/requirements.txt
adversarial-robustness-toolbox
absl-py==0.9.0
alembic==1.4.1
antlr-ast==0.8.1
antlr4-python3-runtime==4.7.2
argunparse==0.1.2
astor==0.8.1
asttokens==1.1.15
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
boto3==1.12.32
botocore==1.15.32
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.2.2
colorama==0.4.1
colorlog==4.1.0
ConfigArgParse==1.1
configparser==4.0.2
cycler==0.10.0
databricks-cli==0.9.1
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
docker==4.2.0
docutils==0.15.2
encrypted-config==0.1.2
entrypoints==0.3
Flask==1.1.1
future==0.18.1
gast==0.2.2
gensim==3.8.1
gitdb2==2.0.6
GitPython==3.0.5
google-api-core==1.16.0
google-auth==1.14.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gorilla==0.3.0
graphviz==0.13.2
grpcio==1.28.1
gym==0.15.4
h5py==2.10.0
horast==0.4.0
idna==2.9
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.10.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterthemes==0.20.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lesscpy==0.13.0
lxml==4.4.1
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mlflow==1.7.0
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.2.1
ordered-set==3.1.1
packaging==20.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.3.0
ply==3.11
pptk==0.1.0
prometheus-client==0.7.1
prometheus-flask-exporter==0.13.0
promise==2.2.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psutil==5.6.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycparser==2.20
pydot==1.2.3
pyglet==1.3.2
Pygments==2.5.2
pyparsing==2.4.5
pypiwin32==223
pyrsistent==0.15.6
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==18.1.1
qgrid==1.1.1
qtconsole==4.6.0
querystring-parser==1.2.4
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
schedule==0.6.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
semver==2.9.0
Send2Trash==1.5.0
sentencepiece==0.1.85
Shapely==1.7.0
simplejson==3.17.0
six==1.14.0
sklearn==0.0
smart-open==1.10.0
smmap2==2.0.5
soupsieve==1.9.4
SQLAlchemy==1.3.15
sqlparse==0.3.1
static-typing==0.2.7
tabulate==0.8.6
tb-nightly==1.14.0a20190603
tensorboard==2.1.1
tensorflow==2.0.1
tensorflow-datasets==1.2.0
tensorflow-estimator==2.0.0
tensorflow-estimator-2.0-preview==2.0.0
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
tensorflow-metadata==0.15.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019060501
tf-nightly-2.0-preview==2.0.0.dev20191002
torch==1.4.0+cpu
torchtext==0.5.0
torchvision==0.5.0+cpu
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
transpyle==0.8.0
typed-ast==1.4.0
typed-astunparse==2.1.4
urllib3==1.25.8
version-query==1.0.5
vowpalwabbit==8.8.1
waitress==1.4.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.0.2
zipp==0.6.0
